pyproject.toml
[project]
name = "pact"
version = "0.1.0"
description = "Contract-first multi-agent software engineering"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "J. Andrew McEntire", email = "j.andrew.mcentire@gmail.com"},
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
pact = "pact.cli:main"

[project.optional-dependencies]
llm = [
    "anthropic>=0.40",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pact"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-x -q"
